requirements.txt
streamlit
streamlit-webrtc
pydub
python-dotenv
langchain
langchain-community
langchain-openai
openai
ffmpeg-python